frontend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
deepgram-sdk==4.0.0
groq==0.4.0
pydantic==2.7.0 